updated url parser name + fixed reference calls
