+++ services/categories-service/src/main/java/com/domesticity/categoriesservice/api/HtmlPagesController.java
@@ -1,6 +1,6 @@
 package com.domesticity.categoriesservice.api;
 
-import com.domesticity.categoriesservice.utilities.DbUrlParser;
+import com.domesticity.categoriesservice.utilities.UrlParser;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
@@ -17,16 +17,16 @@ public class HtmlPagesController {
 
     @GetMapping("/")
     public String getHomePage(Model model) {
-        DbUrlParser dbUrlParser = new DbUrlParser(postgresUrl);
+        UrlParser urlParser = new UrlParser(postgresUrl);
 
         model.addAttribute("service_ver", "v1.0");
         model.addAttribute("spring_ver", "v2.5.8");
         model.addAttribute("build_date", "2022.01.16 23:54.00");
         model.addAttribute("commit_sha", "2c3507de1c78cffab0440d5596fca30c1be1dcc3");
         model.addAttribute("repo_type", repositoryType);
-        model.addAttribute("sql_scheme", dbUrlParser.getScheme());
-        model.addAttribute("sql_ip", dbUrlParser.getIP());
-        model.addAttribute("sql_port", dbUrlParser.getPort());
+        model.addAttribute("sql_scheme", urlParser.getScheme());
+        model.addAttribute("sql_ip", urlParser.getIP());
+        model.addAttribute("sql_port", urlParser.getPort());
 
         return "home";
     }
